send screen, tx list component
